refactored directory 2.0
