+++ Components/Profile/Industries.tsx
