+++ pages/Profile/CreateUser.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import API from "../../api/api";
+import API from "../api/api";
 import Link from "next/link";
 
 const CreateUser = () => {
