+++ pages/index.tsx
@@ -2,9 +2,9 @@ import type { NextPage } from 'next'
 import Head from 'next/head'
 import Image from 'next/image'
 import styles from '../styles/Home.module.css'
-import UserProfile from './Pages/Profile/ReadUser'
-import CreateUser from './Pages/Profile/CreateUser'
-import ReadUser from './Pages/Profile/ReadUser'
+import UserProfile from './Profile/ReadUser'
+import CreateUser from './Profile/CreateUser'
+import ReadUser from './Profile/ReadUser'
 
 const Home: NextPage = () => {
   return (
